cogs/requirements.txt
discord.py
aiohttp
python-dotenv
vaderSentiment
